better classifier; rate limits; don't refilter

+++ laion_icons/filter_images.py
@@ -40,15 +40,18 @@ def main():
         type=str,
         default=os.path.join(root_data_dir(), "filtered", "pending"),
     )
-    parser.add_argument('--ar_limit', type=float, default=1.1)
-    parser.add_argument('--threshold', type=float, default=0.5)
-    parser.add_argument('--batch_size', type=int, default=32)
+    parser.add_argument("--ar_limit", type=float, default=1.1)
+    parser.add_argument("--threshold", type=float, default=0.5)
+    parser.add_argument("--batch_size", type=int, default=32)
     args = parser.parse_args()
 
     os.makedirs(args.output_dir, exist_ok=True)
+    check_dirs = [
+        os.path.join(os.path.dirname(args.output_dir), x) for x in ["success", "failed"]
+    ] + [args.output_dir]
 
-    print('loading classifier...')
-    with open(args.classifier, 'rb') as f:
+    print("loading classifier...")
+    with open(args.classifier, "rb") as f:
         clf = pickle.load(f)
 
     for emb_path in glob.glob(os.path.join(args.emb_dir, "*.npy")):
@@ -62,24 +65,33 @@ def main():
         img_emb = np.load(emb_path)
         df = pq.read_table(meta_path).to_pandas()
         for i in range(0, len(img_emb), args.batch_size):
-            preds = clf.predict_proba(img_emb[i:i+args.batch_size])[:, 1].tolist()
+            preds = clf.predict_proba(img_emb[i : i + args.batch_size])[:, 1].tolist()
             for j, p in enumerate(preds):
+                record = {
+                    k: next(iter(v.values()))
+                    for k, v in df[i + j : i + j + 1].to_dict().items()
+                }
+                if any(
+                    os.path.exists(os.path.join(x, f'{record["key"]}.json'))
+                    for x in check_dirs
+                ):
+                    continue
                 if p >= args.threshold:
-                    record = {k: next(iter(v.values())) for k, v in df[i+j:i+j+1].to_dict().items()}
-                    ar = record['width'] / record['height']
+                    ar = record["width"] / record["height"]
                     if ar < 1:
                         ar = 1 / ar
                     if ar > args.ar_limit:
                         continue
-                    with open(os.path.join(args.output_dir, f'{record["key"]}.json'), 'w') as f:
+                    with open(
+                        os.path.join(args.output_dir, f'{record["key"]}.json'), "w"
+                    ) as f:
                         info = dict(
-                            url=record['url'],
-                            width=record['width'],
-                            height=record['height'],
-                            caption=record['caption'],
+                            url=record["url"],
+                            width=record["width"],
+                            height=record["height"],
+                            caption=record["caption"],
                         )
                         json.dump(info, f)
-
 
 
 if __name__ == "__main__":
